requirements.txt
streamlit>=1.45.1
pytesseract>=0.3.10
Pillow>=9.0.0
numpy>=1.23,<2.0
pandas
opencv-python-headless>=4.5.5.64
openai>=1.0.0
gspread>=5.8.0
google-auth>=2.20.0